Now only keep dataframe rows for downloaded images
There are ~112,000 images originally included in the ChestXRay dataset, but I have only downloaded ~4999 to save space on my local computer. Thus, there were quite a few rows in the data frame corresponding to images that I haven't even downloaded. This is problematic for the data split, as there may be indices in the training split corresponding to images I have not downloaded (I cannot use this in my training set because I have no image to pass to the CNN).

The code I have added now compares the images in the csv to the folder with downloaded images, updating the Dataframe to only keep rows corresponding to images that I have downloaded. With this, I can safely split data knowing that all entries in the train/test/val split will have downloaded images I can pass to a CNN.